Module Octez_smart_rollup_node

module Batcher : sig ... end

module Batcher_events : sig ... end

module Batcher_worker_types : sig ... end

This module contains the parameters for the worker (see Worker) used by the batcher.

module Block_directory_helpers : sig ... end

This module is a helper to convert block ids used in RPCs to useful values.

module Cli : sig ... end

module Commitment_event : sig ... end

This module defines functions that emit the events used for the rollup node when it is storing and publishing commitments (see Commitment).

module Configuration : sig ... end

module Daemon_event : sig ... end

This module defines functions that emit the events used by the smart rollup node daemon (see Daemon).

module Dal_injection_queue : sig ... end

module Event : sig ... end

This module defines functions that emit the events used when the smart rollup node is running (see Daemon).

module Fuel : sig ... end

module Injector : sig ... end

module Interpreter : sig ... end

module Interpreter_event : sig ... end

module Layer1 : sig ... end

This module maintains information about the layer 1 chain.

module Loser_mode : sig ... end

module Metrics : sig ... end

module Node_context : sig ... end

This module describes the execution context of the node.

module Node_context_loader : sig ... end

module Outbox_execution : sig ... end

module Protocol_plugin_sig : sig ... end

module Protocol_plugins : sig ... end

module Publisher : sig ... end

The rollup node stores and publishes commitments for the PVM every `Commitment.sc_rollup_commitment_period` levels.

module Publisher_worker_types : sig ... end

module Purpose : sig ... end

module Pvm_patches : sig ... end

module Pvm_plugin_sig : sig ... end

module Reference : sig ... end

Abstraction module for OCaml references

module Refutation_coordinator : sig ... end

Component for managing refutation games. This module is implemented as a single worker in the rollup node, which takes care of processing new L1 heads, and coordinating the refutation game players. (See Refutation_player).

module Refutation_coordinator_types : sig ... end

module Refutation_game : sig ... end

This module implements the refutation game logic of the rollup node.

module Refutation_game_event : sig ... end

module Refutation_player : sig ... end

module Refutation_player_types : sig ... end

module Repair : sig ... end

module Rollup_node_daemon : sig ... end

module Rollup_node_errors : sig ... end

module Rpc_directory : sig ... end

module Rpc_directory_helpers : sig ... end

module Rpc_server : sig ... end

module Simulation : sig ... end

module Snapshots : sig ... end

module Utils : sig ... end

module Wasm_2_0_0_utilities : sig ... end