Module Batcher_worker_types.Request

type ('a, 'b) t =

| Register : {

order : Z.t option;

messages : string list;

drop_duplicate : bool;

} -> (Octez_smart_rollup.L2_message.id list, Tezos_base.TzPervasives.error Tezos_base.TzPervasives.trace) t
(*
Request to register new L2 messages in the queue. if drop_duplicate is true, then the elements of messages already processed by the batcher, with drop_duplicate = true, are dropped.
*)

| Produce_batches : (unit, Tezos_base.TzPervasives.error Tezos_base.TzPervasives.trace) t
(*
Request to produce new messages batches and submit them to the injector.
*)

| Clear_queues : (unit, Tezos_base.TzPervasives.error Tezos_base.TzPervasives.trace) t
(*
Request to clear all queues, the heap and the batched messages.
*)

| Remove_messages : order_request -> (unit, Tezos_base.TzPervasives.error Tezos_base.TzPervasives.trace) t
(*
Request to remove messages in the heap and batched with order inferior to order_below. If drop_no_order is set, remove also all messages that have no order specified.
*)

Type of requests accepted by the batcher worker.

type view =

| View : (_, _) t -> view

include Tezos_base.Worker_intf.REQUEST with type ('a, 'request_error) t := ('a, 'request_error) t and type view := view

include Tezos_base.Worker_intf.VIEW with type view := view

val encoding : view Data_encoding.t

val pp : Stdlib.Format.formatter -> view -> unit

val view : ('response, 'error) t -> view

The projection function from full request to simple views.