Module Octez_smart_rollup_node.Block_directory_helpers

This module is a helper to convert block ids used in RPCs to useful values.

val block_of_prefix : _ Node_context.t -> Octez_smart_rollup.Rollup_node_services.Arg.block_id -> Tezos_base.TzPervasives.Block_hash.t Tezos_base.TzPervasives.tzresult Lwt.t

Returns the block hash corresponding to a block id.

val block_level_of_id : _ Node_context.t -> Octez_smart_rollup.Rollup_node_services.Arg.block_id -> int32 Tezos_base.TzPervasives.tzresult Lwt.t

Returns the level corresponding to a block id.