Module Octez_smart_rollup_node.Commitment_event

This module defines functions that emit the events used for the rollup node when it is storing and publishing commitments (see Commitment).

val starting : unit -> unit Lwt.t

val stopping : unit -> unit Lwt.t

val section : string list

Section for commitment events.

val last_cemented_commitment_updated : Octez_smart_rollup.Commitment.Hash.t -> int32 -> unit Lwt.t

last_cemented_commitment_updated hash level emits the event that the last cemented commitment was updated to the given hash at the given inbox level.

val last_published_commitment_updated : Octez_smart_rollup.Commitment.Hash.t -> int32 -> unit Lwt.t

last_published_commitment_updated hash level emits the event that the last published commitment was updated to the given hash at the given inbox level.

val compute_commitment : int32 -> unit Lwt.t

compute_commitment level emits the event that a new commitment is being computed and stored for the block at the given level.

val new_commitment : Octez_smart_rollup.Commitment.Hash.t -> int32 -> unit Lwt.t

new_commitment hash level emits the event that a new commitment was built.

val publish_commitment : Octez_smart_rollup.Commitment.Hash.t -> int32 -> unit Lwt.t

publish_commitment hash level emits the event that a new commitment is being published.

val recover_bond : Tezos_crypto.Signature.Public_key_hash.t -> unit Lwt.t

recover_bond staker emits the event that a recover bond operation is being submitted.

val publish_execute_whitelist_update : Octez_smart_rollup.Commitment.Hash.t -> int32 -> int -> unit Lwt.t

publish_execute_whitelist_update hash level index emits the event that a new execute whitelist update is being published.

val execute_outbox_message : outbox_level:int32 -> message_index:int -> Octez_smart_rollup.Outbox_message.summary -> unit Lwt.t

val outbox_message_execution_failed : outbox_level:int32 -> message_index:int -> Octez_smart_rollup.Outbox_message.summary -> Tezos_base.TzPervasives.tztrace -> unit Lwt.t

module Publisher : sig ... end

Events emmitted by the Publisher worker