Module Commitment_event.Publisher

Events emmitted by the Publisher worker

val request_failed : Publisher_worker_types.Request.view -> Tezos_base.Worker_types.request_status -> Tezos_base.TzPervasives.Error_monad.tztrace -> unit Lwt.t

request_failed view status errors emits the event that a worker request view has failed with status after producing errors.

val request_completed : Publisher_worker_types.Request.view -> Tezos_base.Worker_types.request_status -> unit Lwt.t

request_completed view status emits the event that a worker request view has been completed with status.