Module Octez_smart_rollup_node.Configuration

type mode =

| Observer
(*
Only follows the chain and reconstructs inboxes
*)

| Accuser
(*
Only publishes commitments for conflicts and play refutation games
*)

| Bailout

| Batcher
(*
Accept transactions in its queue and batches them on the L1
*)

| Maintenance
(*
Follows the chain and publishes commitments
*)

| Operator
(*
Equivalent to maintenance + batcher
*)

| Custom of Octez_smart_rollup.Operation_kind.t list
(*
In this mode, the system handles only the specific operation kinds defined by the user, allowing for tailored control and flexibility.
*)

Mode for the rollup node

type batcher = {

min_batch_elements : int;
(*
The minimum number elements in a batch for it to be produced when the batcher receives new messages.
*)

min_batch_size : int;
(*
The minimum size in bytes of a batch for it to be produced when the batcher receives new messages.
*)

max_batch_elements : int;
(*
The maximum number of elements that we can put in a batch.
*)

max_batch_size : int option;
(*
The maximum size in bytes of a batch.
*)

}

Configuration for the batcher.

Invariants:

0 < min_batch_size <= max_batch_size <= protocol_max_batch_size

0 < min_batch_elements <= max_batch_elements

type injector = {

retention_period : int;
(*
The number of blocks during which the injector will keep track of an operation (in addition to the confirmation period).
*)

attempts : int;
(*
The number of attempts that will be made to inject an operation.
*)

injection_ttl : int;
(*
The number of blocks after which an operation that is injected but never included is retried.
*)

}

type fee_parameters = Octez_injector.Injector_common.fee_parameter Octez_smart_rollup.Operation_kind.Map.t

type gc_parameters = {

frequency_in_blocks : int32 option;
(*
Frequency at which the GC is triggered.
*)

context_splitting_period : int option;
(*
Number of blocks before splitting the context.
*)

}

type history_mode = Octez_smart_rollup_node_store.Store.State.history_mode =

| Archive
(*
The whole history of the rollup (starting at its genesis) is kept
*)

| Full
(*
Only the history necessary to play refutation games is kept (i.e. after the LCC only)
*)

type outbox_destination_filter =

| Any_destination
(*
Accept any destination.
*)

| Destination_among of string list
(*
Accept destination which match the given list (in base58-check).
*)

Filter on destination of outbox message transactions.

type outbox_entrypoint_filter =

| Any_entrypoint
(*
Accept any entrypoint.
*)

| Entrypoint_among of string list
(*
Accept entrypoints which are listed.
*)

Filter on entrypoints of outbox message transactions.

type outbox_message_filter =

| Transaction of {

destination : outbox_destination_filter;

entrypoint : outbox_entrypoint_filter;

}
(*
Accept transactions which match the filter on their destination and entrypoint.
*)

Filter on outbox messages executed by the rollup node automatically.

type t = {

sc_rollup_address : Tezos_crypto.Hashed.Smart_rollup_address.t;

boot_sector_file : string option;

operators : Purpose.operators;

rpc_addr : string;

rpc_port : int;

acl : Tezos_rpc_http_server.RPC_server.Acl.policy;

metrics_addr : string option;

performance_metrics : bool;

reconnection_delay : float;

fee_parameters : fee_parameters;

mode : mode;

loser_mode : Loser_mode.t;

apply_unsafe_patches : bool;

unsafe_pvm_patches : Pvm_patches.unsafe_patch list;

execute_outbox_messages_filter : outbox_message_filter list;

dal_node_endpoint : Uri.t option;

pre_images_endpoint : Uri.t option;

batcher : batcher;

injector : injector;

l1_blocks_cache_size : int;

l2_blocks_cache_size : int;

prefetch_blocks : int option;

l1_rpc_timeout : float;

loop_retry_delay : float;
(*
Delay in seconds to retry the main loop and the refutation loop after an error.
*)

index_buffer_size : int option;

irmin_cache_size : int option;

log_kernel_debug : bool;

unsafe_disable_wasm_kernel_checks : bool;

no_degraded : bool;

gc_parameters : gc_parameters;

history_mode : history_mode option;

cors : Resto_cohttp.Cors.t;

bail_on_disagree : bool;

}

type Tezos_base.TzPervasives.error +=

| Empty_operation_kinds_for_custom_mode

val history_mode_of_string : string -> history_mode

history_mode_of_string s parses a history_mode from the given string s.

val string_of_history_mode : history_mode -> string

string_of_history_mode p returns a string representation of history_mode p.

val default_data_dir : string

default_data_dir is the default value for data_dir.

val default_storage_dir : string -> string

default_storage_dir returns the default value of the storage dir given a data_dir.

val default_context_dir : string -> string

default_context_dir returns the default value of the directory for persisting the context given a data_dir.

val default_rpc_addr : string

default_rpc_addr is the default value for rpc_addr.

val default_rpc_port : int

default_rpc_port is the default value for rpc_port.

val default_acl : Tezos_rpc_http_server.RPC_server.Acl.policy

default_acl is the default value for acl.

val default_metrics_port : int

default_metrics_port is the default port for the metrics server.

val default_reconnection_delay : float

default_reconnection_delay is the default value for reconnection_delay.

val default_fee_parameter : Octez_smart_rollup.Operation_kind.t -> Octez_injector.Injector_common.fee_parameter

default_fee_parameter operation_kind is the default fee parameter to inject operation on L1. If operation_kind is provided, it returns the default fee parameter for this kind of operation.

val default_fee_parameters : fee_parameters

default_fee_parameters is the default fee parameters configuration build with default_fee_parameter for all purposes.

val default_batcher : batcher

default_batcher is the default configuration parameters for the batcher.

val default_injector : injector

default_injector is the default configuration parameters for the injector.

val default_l1_blocks_cache_size : int

default_l1_blocks_cache_size is the default number of L1 blocks that are cached by the rollup node

val default_l2_blocks_cache_size : int

default_l2_blocks_cache_size is the default number of L2 blocks that are cached by the rollup node

val default_l1_rpc_timeout : float

Default timeout for RPCs to the L1 node.

val default_gc_parameters : gc_parameters

val default_history_mode : history_mode

default_history_mode is the default history mode for the rollup node (Full).

val default_execute_outbox_filter : outbox_message_filter list

Default filter for executing outbox messages is only whitelist updates.

val history_mode_encoding : history_mode Tezos_base.TzPervasives.Data_encoding.t

val max_injector_retention_period : int

max_injector_retention_period is the maximum allowed value for injector_retention_period.

val modes : mode list

This is the list of available modes.

val string_of_mode : mode -> string

string_of_mode mode returns a string representation of the mode specified by the argument mode.

val mode_of_string : string -> mode Tezos_base.TzPervasives.tzresult

mode_of_string s returns the mode represented by string s if it exists.

val description_of_mode : mode -> string

description_of_mode m returns a textual description of the mode m.

val config_filename : data_dir:string -> string

config_filename data_dir returns the configration filename from the data_dir

val purposes_of_mode : mode -> Purpose.ex_purpose list

purposes_of_mode mode returns purposes associated with the provided mode.

val operation_kinds_of_mode : mode -> Octez_smart_rollup.Operation_kind.t list

operation_kinds_of_mode mode returns operation kinds with the provided mode.

val can_inject : mode -> Octez_smart_rollup.Operation_kind.t -> bool

can_inject mode op_kind determines if a given operation kind can be injected based on the configuration settings.

val purposes_matches_mode : mode -> 'kind Purpose.t list -> bool

purposes_matches_mode mode purposes returns true if and only if the given mode supports the given purposes.

val refutation_player_buffer_levels : int

Number of levels the refutation player waits until trying to play for a game state it already played before.

val default_irmin_cache_size : int

val default_index_buffer_size : int

The `default_index_buffer_size` defines the maximum amount of memory reserved for caching index entries before they are written to disk. Essentially, this cache aids the efficiency of the index. The total cache capacity is determined by `index_buffer_size * entry`, with each `entry` occupying approximately 56 bytes. An `entry` represents a single log record which can encompass various details such as a timestamp, message content, severity level, etc.

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

Encoding for configuration.

val encoding_no_default : t Tezos_base.TzPervasives.Data_encoding.t

Encoding for configuration without any default value.

val save : force:bool -> data_dir:string -> t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

save ~force ~data_dir configuration writes the configuration file in data_dir. If force is true, existing configurations are overwritten.

val load : data_dir:string -> t Tezos_base.TzPervasives.tzresult Lwt.t

load ~data_dir loads a configuration stored in data_dir.

module Cli : sig ... end