Module Octez_smart_rollup_node.Daemon_event

This module defines functions that emit the events used by the smart rollup node daemon (see Daemon).

val head_processing : Tezos_base.TzPervasives.Block_hash.t -> int32 -> unit Lwt.t

head_processing hash level emits the event that the block of the given hash and at the given level is being processed.

val new_head_processed : Tezos_base.TzPervasives.Block_hash.t -> int32 -> Ptime.Span.t -> unit Lwt.t

new_head_processed hash level process_time emits the event that the daemon has finished processing the head of the given hash and at the given level in process_time seconds.

val new_head_degraded : Tezos_base.TzPervasives.Block_hash.t -> int32 -> unit Lwt.t

new_head_degraded hash level emits an error event that indicates the rollup node is running in degraded modes and sees an L1 block hash and at the given level.

val processing_heads_iteration : Layer1.head list -> unit Lwt.t

processing_heads_iteration heads emits the event that the heads are going to be processed.

val new_heads_processed : Layer1.head list -> unit Lwt.t

new_heads_processed heads emits the event that the heads were processed.

val new_heads_side_process_finished : Layer1.head list -> unit Lwt.t

new_heads_side_process_finished heads emits the event that the heads side process (e.g. batcher worker) were processed.

val included_operation : ?errors:Tezos_base.TzPervasives.tztrace -> [ `Applied | `Backtracked | `Failed | `Skipped ] -> Octez_smart_rollup.L1_operation.t -> unit Lwt.t

included_operation op result emits an event that an operation for the rollup was included in a block.

val migration : catching_up:bool -> (Tezos_base.TzPervasives.Protocol_hash.t * int) -> (Tezos_base.TzPervasives.Protocol_hash.t * int) -> unit Lwt.t

migration ~catching_up (old_protocol, old_protocol_level) (new_protocol, new_protocol_level) emits and event for when the rollup node detects and handles a protocol migration.

val switched_protocol : Tezos_base.TzPervasives.Protocol_hash.t -> int -> Octez_smart_rollup.Rollup_constants.protocol_constants -> unit Lwt.t

val error : Tezos_base.TzPervasives.tztrace -> unit Lwt.t

Emit a fatal error for the daemon.

val degraded_mode : unit -> unit Lwt.t

Emit an event for when the node enters the degraded mode to only play refutations.

val refutation_loop_retry : float -> unit Lwt.t

val main_loop_retry : float -> unit Lwt.t

val exit_bailout_mode : unit -> unit Lwt.t

Emit an event when the node exits after recovering the operator's stakes.