Module Octez_smart_rollup_node.Dal_injection_queue

val init : _ Node_context.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

init node_ctxt initializes and starts the DAL injection worker.

val register_dal_messages : messages:string list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

register_dal_messages ~messages registers new messages in the queue for future injection on L1 and DAL.

val get_injection_ids : unit -> Injector.Inj_operation.id list Tezos_base.TzPervasives.tzresult

get_injection_ids () returns the current injection IDs known by the DAL injection queue.

val forget_injection_id : Octez_injector.Injector_sigs.Id.t -> unit Tezos_base.TzPervasives.tzresult

forget_injection_id id removes the given injection id and its associated information from the DAL injection queue, if it exists.

val produce_dal_slots : level:int32 -> unit Tezos_base.TzPervasives.tzresult Lwt.t

produce_dal_slots ~level produces and injects DAL slots on top of block at level level from the pending data injected to the DAL injection queue.

val set_dal_slot_indices : Tezos_dal_node_services.Types.slot_index list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

set_dal_slot_indices idx sets the list of slot indices on which the DAL injection worker will publish DAL slots.