Module Octez_smart_rollup_node.Event

This module defines functions that emit the events used when the smart rollup node is running (see Daemon).

val starting_node : unit -> unit Lwt.t

val node_is_ready : rpc_addr:string -> rpc_port:int -> unit Lwt.t

val rollup_exists : addr:Octez_smart_rollup.Address.t -> kind:Octez_smart_rollup.Kind.t -> unit Lwt.t

rollup_exists addr kind emits the event that the smart rollup node is interacting with the rollup at address addr and of the given kind.

val shutdown_node : int -> unit Lwt.t

shutdown_node exit_status emits the event that the smart rollup node is stopping with exit status exit_status.

val starting_metrics_server : host:string -> port:int -> unit Lwt.t

starting_metrics_server ~metrics_addr ~metrics_port emits the event that the metrics server for the rollup node is starting.

val metrics_ended : string -> unit Lwt.t

metrics_ended error emits the event that the metrics server has ended with a failure.

val metrics_ended_dont_wait : string -> unit

metrics_ended error emits the event that the metrics server has ended with a failure. (Doesn't wait for event to be emited.

val metrics_error : string -> unit

metrics_error error emits the event that the metrics has an error.

val kernel_debug : string -> unit Lwt.t

kernel_debug str emits the event that the kernel has logged str.

val simulation_kernel_debug : string -> unit Lwt.t

simulation_kernel_debug str emits the event that the kernel has logged str during a simulation.

val kernel_debug_dont_wait : string -> unit

kernel_debug str emits the event that the kernel has logged str. (Doesn't wait for event to be emitted)

val warn_dal_enabled_no_node : unit -> unit Lwt.t

warn_dal_enabled_no_node () emits a warning for when DAL is enabled in the protocol but the rollup node has no DAL node.

val waiting_first_block : Tezos_base.TzPervasives.Protocol_hash.t -> unit Lwt.t

Emit event that the node is waiting for the first block of its protocol.

val received_first_block : Tezos_base.TzPervasives.Block_hash.t -> Tezos_base.TzPervasives.Protocol_hash.t -> unit Lwt.t

Emit event that the node received the first block of its protocol.

val detected_protocol_migration : unit -> unit Lwt.t

Emit event that the node will shutdown because of protocol migration.

val acquiring_lock : unit -> unit Lwt.t

acquiring_lock () emits an event to indicate that the node is attempting to acquire a lock on the data directory.

val calling_gc : gc_level:int32 -> head_level:int32 -> unit Lwt.t

calling_gc ~gc_level ~head_level emits the event that the GC is started for level gc_level.

val gc_levels_storage_failure : unit -> unit Lwt.t

gc_levels_storage_failure () emits an event which indicates that GC level values (last_gc_level, first_available_level) could not be written to storage.

val convert_history_mode : Configuration.history_mode -> Configuration.history_mode -> unit Lwt.t

convert_history_mode old_history_mode new_history_mode emits an event for when the history mode of the rollup node is changed.

val gc_finished : gc_level:int32 -> head_level:int32 -> unit Lwt.t

gc_finished ~gc_level ~head_level emits the event that the GC is finished for level gc_level.