Module Fuel.Free

Free fuel where consumption has no effect.

type t

val consume : t -> t -> t option

consume consumption fuel consumes the consumption amount from the original fuel. It returns None when the consumption is greater than the original fuel or Some remaining_fuel.

val one_tick_consumption : t

The amount of fuel required to run one PVM tick.

one_tick_consumption = of_ticks 1L

val of_ticks : int64 -> t

of_ticks ticks gives the amount of fuel required to execute the amount of ticks.

val is_empty : t -> bool

val max_ticks : t -> int64

The maximum number of ticks that can be executed with the given amount of fuel.

max_ticks ∘ of_ticks = Fun.id
of_ticks ∘ max_ticks = Fun.id