Module Octez_smart_rollup_node.Interpreter_event

module Simple : sig ... end

val transitioned_pvm : int32 -> Octez_smart_rollup.State_hash.t -> Z.t -> int -> unit Lwt.t

transition_pvm inbox_level hash tick n emits the event that a PVM transition is leading to the state of the given hash by processing n messages at tick.

val intended_failure : level:int -> message_index:int -> message_tick:int64 -> internal:bool -> unit Lwt.t

intended_failure level message_index message_tick internal emits the event that an intended failure has been injected at some given level, during the processing of a given message_index and at tick message_tick during this message processing. internal is true if the failure is injected in a PVM internal step. internal is false if the failure is injected in the input to the PVM.

val missing_pre_image : hash:string -> unit Lwt.t

val fetched_incorrect_pre_image : expected_hash:string -> content_hash:string -> unit Lwt.t

val patching_genesis_state : Pvm_patches.unsafe_patch -> unit Lwt.t

val fast_exec_panic : exn -> unit Lwt.t