Module Octez_smart_rollup_node.Loser_mode

type t

A list of failures.

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

val no_failures : t

no_failures are planned.

val make : string -> t option

make s parses a list of integers separated by spaces that is a periodic sequence of triple level message_index message_tick representing a failure that the rollup node is supposed to make. This function returns None if the input string is not syntactically correct.

val is_failure : t -> level:int -> message_index:int -> int64 list

is_failure failures ~level ~message_index returns message_ticks where a failure is supposed to happen at the point of the rollup node processing of a given inbox level, a given message_index and for all message_ticks. Ticks are sorted by increasing order.