Module Octez_smart_rollup_node.Metrics

val sc_rollup_node_registry : Prometheus.CollectorRegistry.t

The collector registry for the rollup node metrics.

val active_metrics : Configuration.t -> unit

Enables the wrapper variable

val wrap : (unit -> unit) -> unit

Wrap a function to be processed if a metrics address is set up in the configuration.

val wrap_lwt : (unit -> (unit, 'error) Stdlib.result Lwt.t) -> (unit, 'error) Stdlib.result Lwt.t

val metrics_serve : string option -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

metrics_server metrics_addr runs a server for the rollup metrics on metrics_addr. The metrics are accessible thanks to a /metrics request.

val print_csv_metrics : Stdlib.Format.formatter -> 'a Prometheus.MetricFamilyMap.t -> unit

print_csv_metrics ppf metrics prints the metrics as CSV.

module Refutation : sig ... end

module Info : sig ... end

The node info metrics

module Inbox : sig ... end

The metrics related to Inboxes

module Batcher : sig ... end

module DAL_batcher : sig ... end

module GC : sig ... end

module Performance : sig ... end