Module Metrics.DAL_batcher

val set_dal_batcher_queue_length : int -> unit

Set the number of messages in the queue waiting to be published on the DAL

val set_dal_injections_queue_length : int -> unit

Set the number of DAL slots which were recently published and not yet forgotten