Module Metrics.Info

The node info metrics

val init_rollup_node_info : Configuration.t -> genesis_level:int32 -> genesis_hash:Octez_smart_rollup.Commitment.Hash.t -> pvm_kind:string -> history_mode:Configuration.history_mode -> unit

Initializes the metric for rollup info with a the given arguments as label values

val set_lcc_level_local : int32 -> unit

Set the last cemented commitment level available locally

val set_lcc_level_l1 : int32 -> unit

Set the last cemented commitment level seen on the network

val set_lpc_level_local : int32 -> unit

Set the last published commitment level available locally

val set_lpc_level_l1 : int32 -> unit

Set the last published commitment level seen on the network

val set_proto_info : Tezos_base.TzPervasives.Protocol_hash.t -> Octez_smart_rollup.Rollup_constants.protocol_constants -> unit

Set the protocol specific information for the current protocol