Module Metrics.Refutation

type state =

| OurTurn

| TheirTurn

| Timeout

val set_number_of_conflict : int -> unit

Set the number of current conflict for this rollup node

val set_state_refutation_game : ?labels:string list -> state -> unit

Set the state of a refutation game whether it's our turn or the opponent

val set_block_timeout : ?labels:string list -> int -> unit

Set the number of block before the player timeout in the game

val clear_state_refutation_game : string list -> unit

Clear the state of a refutation game