Module Node_context.Internal_for_tests

val write_protocols_in_store : [> `Write ] store -> Octez_smart_rollup_node_store.Store.Protocols.proto_info list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val unsafe_get_store : 'a t -> 'a Octez_smart_rollup_node_store.Store.t

Extract the underlying store from the node context. This function is unsafe to use outside of tests as it breaks the abstraction barrier provided by the Node_context.