Module Node_context.Node_store

Exposed functions to manipulate Node_context store outside of this module

val init_with_migration : 'a Tezos_layer2_store.Store_sigs.mode -> Octez_smart_rollup.Metadata.t -> data_dir:string -> 'a store Tezos_base.TzPervasives.tzresult Lwt.t

val close : 'a store -> unit Lwt.t

close store closes the store

val check_and_set_history_mode : 'a Tezos_layer2_store.Store_sigs.mode -> 'a store -> Configuration.history_mode option -> unit Tezos_base.TzPervasives.tzresult Lwt.t

check_and_set_history_mode store history_mode checks the compatibility between given history mode and that of the store. History mode can be converted from Archive to Full. Trying to convert from Full to Archive will trigger an error.

val of_store : 'a Octez_smart_rollup_node_store.Store.t -> 'a store