Module Node_context_loader.For_snapshots

val create_node_context : Tezos_client_base.Client_context.full -> Node_context.current_protocol -> [< `Read | `Write Read ] as 'a Octez_smart_rollup_node_store.Store.t -> 'a Tezos_layer2_store.Context.t -> data_dir:string -> apply_unsafe_patches:bool -> 'a Node_context.t Tezos_base.TzPervasives.tzresult Lwt.t

create_node_context cctxt protocol store context ~data_dir creates a node context which does not monitor the L1 chain but which can be used to reconstruct the context from historical data. This function is used by the Snapshots module.