Module Octez_smart_rollup_node.Outbox_execution

val publish_execute_whitelist_update_message : Node_context.rw -> unit Tezos_base.TzPervasives.tzresult Lwt.t

publish_execute_whitelist_update_message node_ctxt checks, starting from the lcc, for the presence of the whitelist update message to execute. It publishes and executes the most recent/closest one.

val publish_executable_messages : Node_context.rw -> unit Tezos_base.TzPervasives.tzresult Lwt.t

publish_executable_message node_ctxt publishes on L1 all outbox messages that can be executed and that have not yet been executed.