Module Octez_smart_rollup_node.Protocol_plugin_sig

module type RPC_DIRECTORY = sig ... end

Protocol specific RPC directory, used as a dynamic directory in the protocol agnostic rollup node.

module type DAL_SLOTS_TRACKER = sig ... end

Protocol specific functions to track attested DAL slots of L1 blocks.

module type INBOX = sig ... end

Protocol specific functions to reconstruct inboxes from L1 blocks.

module type BATCHER_CONSTANTS = sig ... end

Protocol specific constants for the batcher.

module type LAYER1_HELPERS = sig ... end

Protocol specific functions to interact with the L1 node.

module type L1_PROCESSING = sig ... end

Protocol specific functions for processing L1 blocks.

module type PARTIAL = sig ... end

Partial protocol plugin with just the PVM and the function to access the Layer1. This signature exists in order to build plugins for the interpreter and the refutation games while avoiding circular dependencies.

module type REFUTATION_GAME_HELPERS = sig ... end

Protocol specific refutation helper functions.

module type S = sig ... end

Signature of protocol plugins for the rollup node. NOTE: the plugins have to be registered to be made available to the rollup node.