Module type Protocol_plugin_sig.L1_PROCESSING

Protocol specific functions for processing L1 blocks.

val process_l1_block_operations : catching_up:bool -> Node_context.rw -> Layer1.header -> unit Tezos_base.TzPervasives.tzresult Lwt.t

React to L1 operations included in a block of the chain. When catching_up is true, the process block is in the past and the failure condition of the process differs (e.g. if it detects a whitelist update execution, it does not fail when the operator is not in the whitelist).