Module type Protocol_plugin_sig.PARTIAL

Partial protocol plugin with just the PVM and the function to access the Layer1. This signature exists in order to build plugins for the interpreter and the refutation games while avoiding circular dependencies.

val protocol : Tezos_base.TzPervasives.Protocol_hash.t

The protocol for which this plugin is.

module Layer1_helpers : LAYER1_HELPERS

module Pvm : Pvm_plugin_sig.S