Module PARTIAL.Layer1_helpers

val prefetch_tezos_blocks : Layer1.t -> Layer1.head list -> unit

prefetch_tezos_blocks l1_ctxt blocks prefetches the blocks asynchronously. NOTE: the number of blocks to prefetch must not be greater than the size of the blocks cache otherwise they will be lost.

val get_last_cemented_commitment : Tezos_client_base.Client_context.full -> Octez_smart_rollup.Address.t -> Node_context.lcc Tezos_base.TzPervasives.tzresult Lwt.t

val get_last_published_commitment : ?allow_unstake:bool -> Tezos_client_base.Client_context.full -> Octez_smart_rollup.Address.t -> Tezos_crypto.Signature.public_key_hash -> Octez_smart_rollup.Commitment.t option Tezos_base.TzPervasives.tzresult Lwt.t

val get_kind : Tezos_client_base.Client_context.full -> Octez_smart_rollup.Address.t -> Octez_smart_rollup.Kind.t Tezos_base.TzPervasives.tzresult Lwt.t

val genesis_inbox : Tezos_client_base.Client_context.full -> genesis_level:int32 -> Octez_smart_rollup.Inbox.t Tezos_base.TzPervasives.tzresult Lwt.t

val retrieve_constants : ?block:Tezos_shell_services.Block_services.block -> Tezos_client_base.Client_context.full -> Octez_smart_rollup.Rollup_constants.protocol_constants Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve protocol agnotic constants for the head of the chain.

val retrieve_genesis_info : Tezos_client_base.Client_context.full -> Octez_smart_rollup.Address.t -> Node_context.genesis_info Tezos_base.TzPervasives.tzresult Lwt.t

val get_boot_sector : Tezos_base.TzPervasives.Block_hash.t -> _ Node_context.t -> string Tezos_base.TzPervasives.tzresult Lwt.t

get_boot_sector block_hash node_ctxt retrieves the boot sector from the rollup origination operation in block block_hash. Precondition: block_hash has to be the block where the rollup was originated.

val find_whitelist : Tezos_client_base.Client_context.full -> ?block:Tezos_base.TzPervasives.Block_hash.t -> Octez_smart_rollup.Address.t -> Tezos_crypto.Signature.public_key_hash list option Tezos_base.TzPervasives.tzresult Lwt.t

Find and retrieve the whitelist the rollup at a given block (if provided) or the head.

val find_last_whitelist_update : Tezos_client_base.Client_context.full -> Octez_smart_rollup.Address.t -> (Z.t * Stdlib.Int32.t) option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve information about the last whitelist update on L1.

val get_commitment : Tezos_client_base.Client_context.full -> Octez_smart_rollup.Address.t -> Octez_smart_rollup.Commitment.Hash.t -> Octez_smart_rollup.Commitment.t Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve a commitment published on L1.

val get_balance_mutez : Tezos_client_base.Client_context.full -> ?block:Tezos_base.TzPervasives.Block_hash.t -> Tezos_crypto.Signature.public_key_hash -> int64 Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve the balance in mutez for a given public key hash.