Module Fueled.Accounted

val eval_block_inbox : fuel:Fuel.Accounted.t -> _ Node_context.t -> (Octez_smart_rollup.Inbox.t * string list) -> Tezos_layer2_store.Context.pvmstate -> Fuel.Accounted.t Pvm_plugin_sig.eval_result Tezos_base.TzPervasives.tzresult Lwt.t

eval_block_inbox ~fuel node_ctxt (inbox, messages) state evaluates the messages for the inbox in the given state of the PVM and returns the evaluation result containing the new state, the number of messages, the inbox level and the remaining fuel.

val eval_messages : ?reveal_map:string Utils.Reveal_hash_map.t -> _ Node_context.t -> Fuel.Accounted.t Pvm_plugin_sig.eval_state -> Fuel.Accounted.t Pvm_plugin_sig.eval_result Tezos_base.TzPervasives.tzresult Lwt.t

eval_messages ?reveal_map ~fuel node_ctxt ~message_counter_offset state inbox_level messages evaluates the messages for inbox level inbox_level in the given state of the PVM and returns the evaluation results containing the new state, the remaining fuel, and the number of ticks for the evaluation of these messages. If messages is empty, the PVM progresses until the next input request (within the allocated fuel). message_counter_offset is used when we evaluate partial inboxes, such as during simulation. When reveal_map is provided, it is used as an additional source of data for revelation ticks.