Module Octez_smart_rollup_node.Pvm_patches

type unsafe_patch =

| Increase_max_nb_ticks of int64
(*
Increase the maximum number of ticks.
*)

| Patch_durable_storage of {

key : string;

value : string;

}

The type of individual unsafe patch content.

type kind =

| Hardcoded
(*
Hardcoded by the rollup node
*)

| User_provided
(*
Provided by the user in the configuration file
*)

Where the patch comes from

type t = private (unsafe_patch * kind) list

The type of registered patches for the PVM.

val unsafe_patch_encoding : unsafe_patch Tezos_base.TzPervasives.Data_encoding.t

Encoding for unsafe patches.

val pp_unsafe_patch : Stdlib.Format.formatter -> unsafe_patch -> unit

Pretty printer for unsafe patches.

val make : Octez_smart_rollup.Kind.t -> Octez_smart_rollup.Address.t -> unsafe_patch list -> t Tezos_base.TzPervasives.tzresult

make kind address patches builds the patches from the user provided list patches and adds the hardcoded PVM patches for the rollup address.