Module Octez_smart_rollup_node.Pvm_plugin_sig

type 'fuel eval_state = {

state : Tezos_layer2_store.Context.pvmstate;
(*
The actual PVM state.
*)

state_hash : Octez_smart_rollup.State_hash.t;
(*
Hash of state.
*)

tick : Z.t;
(*
Tick of state.
*)

inbox_level : int32;
(*
Inbox level in which messages are evaluated.
*)

message_counter_offset : int;
(*
Offset for message index, which corresponds to the number of messages of the inbox already evaluated.
*)

remaining_fuel : 'fuel;
(*
Fuel remaining for the evaluation of the inbox.
*)

remaining_messages : string list;
(*
Messages of the inbox that remain to be evaluated.
*)

}

Evaluation state for the PVM.

type 'fuel eval_result = {

state : 'fuel eval_state;

num_ticks : Z.t;

num_messages : int;

}

Evaluation result for the PVM which contains the evaluation state and additional information.

module Tick_state_cache : Aches_lwt.Lache.MAP_RESULT with type key = Z.t

type state_cache = (Fuel.Accounted.t eval_state, Tezos_base.TzPervasives.tztrace) Tick_state_cache.t

val make_state_cache : int -> ('a, 'b) Tick_state_cache.t

module type FUELED_PVM = sig ... end

module type S = sig ... end