Module Octez_smart_rollup_node.Refutation_coordinator

Component for managing refutation games. This module is implemented as a single worker in the rollup node, which takes care of processing new L1 heads, and coordinating the refutation game players. (See Refutation_player).

val init : Node_context.rw -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Initiatilize the refuation coordinator, if the rollup node mode supports it.

val start_in_mode : Configuration.mode -> bool

Whether the refutation coordinator is run in the given mode.

val process : Layer1.head -> unit Tezos_base.TzPervasives.tzresult Lwt.t

process head processes a new l1 head. This means that the coordinator will:

Gather all existing conflicts

Launch new refutation players for each conflict concerning the operator that doesn't have a player in this node

Kill all players whose conflict has disappeared from L1

Make all players play a step in the refutation

val shutdown : unit -> unit Lwt.t

Shutdown the refutation coordinator.

val worker_status : unit -> [ `Running | `Not_running | `Crashed of exn ]

Returns the status of the publisher worker.