Module Octez_smart_rollup_node.Refutation_game

This module implements the refutation game logic of the rollup node.

val play_opening_move : [< `Read | `Write Read ] Node_context.t -> Octez_smart_rollup.Game.conflict -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

play_opening_move node_ctxt conflict injects the opening refutation game move for conflict.

val play : Node_context.rw -> Pvm_plugin_sig.state_cache -> self:Tezos_crypto.Signature.public_key_hash -> commitment_period_tick_offset:Z.t -> Octez_smart_rollup.Game.t -> Tezos_crypto.Signature.public_key_hash -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

play head_block plugin node_ctxt cache ~self ~commitment_period_tick_offset game opponent injects the next move in the refutation game played by self and opponent, where commitment_period_tick_offset is the tick offset for the commitment period for the conflict.