Parameter Worker.Request

type ('response, 'error) t

The type of events. It is possible to wait for an event to be processed from outside the worker using push_request_and_wait. In this case, the handler for this event can return a value. The parameter is the type of this value.

include Tezos_base.Worker_intf.VIEW

type view

val encoding : view Data_encoding.t

val pp : Stdlib.Format.formatter -> view -> unit

val view : ('response, 'error) t -> view

The projection function from full request to simple views.