Module Octez_smart_rollup_node.Rpc_directory_helpers

module type PARAM = sig ... end

This module defines subcontext type of the subdirectory and the way to project it from Node_context and a path prefix.

module type PARAM_PREFIX = sig ... end

Parameter signature to use with Make_directory, where the prefix is given.

module Make_sub_directory (S : PARAM) : sig ... end

This module is a helper to register your endpoints and build a resulting subdirectory eventually.

module Make_directory (S : PARAM_PREFIX) : sig ... end

This module is a helper to register your endpoints and build a resulting top level directory eventually.