Module Rpc_directory_helpers.Make_directory

This module is a helper to register your endpoints and build a resulting top level directory eventually.

Parameters

module S : PARAM_PREFIX

Signature

include sig ... end

val register0 : ([< Resto.meth ], 'prefix, 'prefix, 'query, 'input, 'output) Tezos_rpc.Service.t -> (S.subcontext -> 'query -> 'input -> 'output Tezos_base.TzPervasives.tzresult Lwt.t) -> unit

Register an endpoint with no parameters in the path.

val register1 : ([< Resto.meth ], 'prefix, 'prefix * 'param1, 'query, 'input, 'output) Tezos_rpc.Service.t -> (S.subcontext -> 'param1 -> 'query -> 'input -> 'output Tezos_base.TzPervasives.tzresult Lwt.t) -> unit

Register an endpoint with a single parameter in the path.

val gen_register0 : ([< Resto.meth ], 'prefix, 'prefix, 'query, 'input, 'output) Tezos_rpc.Service.t -> (S.subcontext -> 'query -> 'input -> 'output Tezos_rpc.Answer.t Lwt.t) -> unit

Register an endpoint that specifies how to process incoming queries and inputs; this function is intended for handling asynchronous contexts.

val build_sub_directory : S.context -> S.prefix Tezos_rpc.Directory.t

Build sub-directory with registered endpoints with respect to Node_context.

val of_subdirectory : S.prefix Tezos_rpc.Directory.t -> unit Tezos_rpc.Directory.t

Build a top-level directory from a sub-directory by prefixing it with prefix.

val build_directory : S.context -> unit Tezos_rpc.Directory.t

Build top-level directory with registered endpoints with respect to Node_context.