Module Octez_smart_rollup_node.Rpc_server

type t

module Acl : sig ... end

val start : rpc_addr:string -> rpc_port:int -> acl:Tezos_rpc_http_server.RPC_server.Acl.policy -> cors:Resto_cohttp.Cors.t -> unit Tezos_rpc.Directory.t -> t Tezos_base.TzPervasives.tzresult Lwt.t

start ~rpc_addr ~rpc_port ~acl ~cors node_ctxt config starts an RPC server listening for requests on the port rpc_port and address rpc_addr.

val shutdown : t -> unit Lwt.t

Shutdown a running RPC server. When this function is called, the rollup node will stop listening to incoming requests.