Module Octez_smart_rollup_node.Wasm_2_0_0_utilities

val dump_durable_storage : block:Tezos_shell_services.Block_services.block -> data_dir:string -> file:string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

dump_durable_storage ~block ~data_dir ~file writes to file the current state of the WASM PVM from data_dir, that is the state of the WASM PVM for the given block.

val patch_durable_storage : data_dir:string -> key:string -> value:string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

patch_durable_storage ~data_dir ~key ~value adds value to key into the durable storage of PVM state. If the value exists it's replaced.

val hooks : check_invalid_kernel:bool -> Tezos_scoru_wasm.Hooks.t

Hooks to be used for the WASM PVM.